Change format identifier of SPI_processed to uint64.

+++ lib/repack.c
@@ -1306,8 +1306,8 @@ repack_index_swap(PG_FUNCTION_ARGS)
 					 orig_idx_oid);
 	execute(SPI_OK_SELECT, str.data);
 	if (SPI_processed != 1)
-		elog(ERROR, "Could not find index 'index_%u', found %d matches",
-			 orig_idx_oid, SPI_processed);
+		elog(ERROR, "Could not find index 'index_%u', found %lu matches",
+			 orig_idx_oid, (uint64) SPI_processed);
 
 	tuptable = SPI_tuptable;
 	desc = tuptable->tupdesc;
